backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "teamified-backend"
version = "1.0.0"
description = "Teamified Candidates Portal - Backend API"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]>=0.24",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.12",
    "greenlet>=3.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "email-validator>=2.1",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx>=0.25",
    "structlog>=23.2",
    "langchain>=0.1.0",
    "langchain-community>=0.4.0",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "websockets>=12.0",
    "aiohttp>=3.13.2",
    "aiofiles>=25.1.0",
    "supabase>=2.23.0",
    "pgvector>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-mock",
    "pytest-cov",
    "ruff",
    "black",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
]
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | \.mypy_cache
)/
'''

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    ".venv",
    "venv",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --strict-markers --cov=app --cov-report=term-missing"

[tool.coverage.run]
source = ["app"]
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
